requirements.txt
uvicorn~=0.25.0
fastapi~=0.108.0
jinja2~=3.1.2
pony~=0.7.17
pydantic~=2.5.3
python-jose~=3.3.0